refactor: split loader and ui modules
